requirements.txt
httpx
beautifulsoup4
lxml
google-genai
python-dotenv
apscheduler
pandas
plotly
streamlit
pytest
